requirements.txt
bcrypt==4.0.1
blinker==1.6.2
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.7
colorama==0.4.6
cryptography==41.0.4
dnspython==2.4.2
Flask==2.3.3
Flask-Cors==4.0.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
jwt==1.3.1
MarkupSafe==2.1.3
mongoengine==0.27.0
pycparser==2.21
PyJWT==2.8.0
pymongo==4.5.0
python-dotenv==1.0.0
razorpay==1.4.1
requests==2.31.0
urllib3==2.0.5
Werkzeug==2.3.7
